Instructions for Adding Mugshot Transitions
Added instructions in include/constants/battle_transition.h for how to
add custom mugshot battle transitions. Also included a "template" for
the palettes used for the mugshot backgrounds.

An important note with the palettes: They must have a newline at the end
of the file AND have line endings of CRLF.