Custom Wild Battles and Shiny Rate Script Command
Begin implementation of function that allows for "boss" Pokemon to be
implemented that can be triggered through map scriptings. Done through
the StartWildBossBattle function (right now has a temporary Zigzagoon
for testing).

Also implemented new script macro setextrashinytries
(SetExtraShinyTries) that takes a u16 value and stores it in a new
global variable gExtraShinyTries. This new global variable takes the
place of shinyRolls that was in the function CreateBoxMon, and it is
used to indicate the number of extra attempts to be made for generating
a shiny Pokemon.

The global variable gUnusedFirstBattleVar1 in battle.h was converted
from a u32 to a u16 to allow for gExtraShinyTries to be implemented
without taking up more space.