Reversion and Deletion of Maps
I felt as though the direction the mod was taking wasn't what I
originally intended. My primary focus is on exploration and Pokedex
completion, but I was gearing up to make an entire post-game story.
Right now, that's not the idea. Maybe in the future I'll reimplement
some ideas I came up with, but for now, I'll be focusing on expanding
Hoenn to make it more interesting from an exploration perspective.